Vladimirov plots, documentation
